requirements.txt
# Core dependencies
accelerate==1.7.0
aiofiles==24.1.0
aiohttp==3.12.11
bitsandbytes==0.46.0
dataclasses-json==0.6.7
datasets==3.6.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.13.1
frozenlist==1.6.2
fsspec==2025.3.0
gradio==5.33.0
gradio_client==1.10.2
huggingface-hub==0.32.4
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-text-splitters==0.3.8
markdown-it-py==3.0.0
multidict==6.4.4
numpy==2.2.6
openai==1.30.1
orjson==3.10.18
packaging==24.2
peft==0.15.2
PyPDF2==3.0.1
python-dotenv==1.1.0
python-multipart==0.0.20
regex==2024.11.6
requests==2.31.0
safetensors==0.5.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
streamlit==1.45.1
tenacity==9.1.2
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
transformers==4.52.4
triton==3.3.0
typer==0.16.0
typing_extensions==4.11.0
uvicorn==0.34.3
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0